--- Sources/Utilities/NSLocalizedString+Extension.swift
@@ -1,8 +1,0 @@
-//
-//  NSLocalizedString+Extension.swift
-//  ApplozicSwift
-//
-//  Created by Shivam Pokhriyal on 08/10/18.
-//
-
-import Foundation
